+++ app/scripts/services/dimStoreService.factory.js
@@ -40,9 +40,9 @@
       // How much of items like this item can fit in this store?
       capacityForItem: function(item) {
         if (!item.bucket) {
-          throw new Error("item needs a 'bucket' field");
+          console.error("item needs a 'bucket' field", item);
         }
-        return bucketSizes[item.bucket];
+        return bucketSizes[item.bucket] || 10;
       },
       // How many *more* items like this item can fit in this store?
       spaceLeftForItem: function(item) {
@@ -213,7 +213,7 @@
             var store;
             var items = [];
             if (!raw) {
-              return;
+              return undefined;
             }
 
             if (raw.id === 'vault') {
@@ -346,37 +346,59 @@
     function processSingleItem(definitions, itemBucketDef, statDef, objectiveDef, perkDefs, talentDefs, yearsDefs, progressDefs, item) {
       var itemDef = definitions[item.itemHash];
       // Missing definition?
-      if (itemDef === undefined || itemDef.itemName === 'Classified') {
+      if (!itemDef || itemDef.itemName === 'Classified') {
         // maybe it is classified...
         itemDef = {
           classified: true,
           icon: '/img/misc/missing_icon.png'
         };
 
+        if (item.itemHash === 194424271) {
+          itemType = 'Armor Shader';
+          itemDef.bucketTypeHash = 2973005342;
+          itemDef.classType = 3;
+          itemDef.itemTypeName = 'Armor Shader';
+          itemDef.description = '';
+          itemDef.itemName = 'Walkabout - Classified';
+          itemDef.nonTransferrable = true;
+          itemDef.equipRequiredLevel = 0;
+          itemDef.equipment = true;
+          item.isEquipment = true;
+        }
+
+        if (item.itemHash === 1963806104) {
+          itemType = 'Mystery Bag';
+          itemDef.bucketTypeHash = 1469714392;
+          itemDef.icon = "/common/destiny_content/icons/3651da8a8b0add3161e840c7104078ed.jpg";
+          itemDef.classType = 3;
+          itemDef.itemTypeName = itemType;
+          itemDef.description = "Contains 1 Guaranteed Item and up to 4 Possible Items. This item is nonreturnable.";
+          itemDef.itemName = "Sterling Treasure";
+          itemDef.maxStackSize = 99;
+          itemDef.nonTransferrable = true;
+          itemDef.equipRequiredLevel = 0;
+          itemDef.equipment = false;
+          item.isEquipment = false;
+        }
+
         // unidentified item.
-        if(!itemDef.itemName) {
+        if (!itemDef.itemName) {
           console.warn('Missing Item Definition:\n\n', item, '\n\nplease contact a developer to get this item added.');
           window.onerror("Missing Item Definition - " + JSON.stringify(_.pick(item, 'canEquip', 'cannotEquipReason', 'equipRequiredLevel', 'isEquipment', 'itemHash', 'location', 'stackSize', 'talentGridHash')), 'dimStoreService.factory.js', 491, 11);
         }
       }
 
-      if (_.isUndefined(itemDef.itemTypeName) || _.isUndefined(itemDef.itemName)) {
+      if (!itemDef.itemTypeName || !itemDef.itemName) {
         return null;
       }
 
       console.log('get item type', item, itemDef, itemBucketDef)
       var itemType = getItemType(item, itemDef, itemBucketDef);
-
-      if (item.itemHash === 937555249) {
-        itemType = "Material";
-      }
-
-      var weaponClass = null, weaponClassName = null;
-
-
       if (!itemType) {
         return null;
       }
+
+      var weaponClass = null, weaponClassName = null;
 
       if (itemType.hasOwnProperty('general') && itemType.general !== '') {
         weaponClass = itemType.weaponClass;
@@ -385,8 +407,7 @@
       }
 
       var itemSort = sortItem(itemDef.itemTypeName);
-
-      if (_.isUndefined(itemSort)) {
+      if (!itemSort) {
         console.log(itemDef.itemTypeName + " does not have a sort property.");
       }
 
@@ -403,18 +424,15 @@
 
       var dmgName = [null, 'kinetic', 'arc', 'solar', 'void'][item.damageType];
 
-      var talentGrid = buildTalentGrid(item, talentDefs, progressDefs, perkDefs),
-          stats = buildStats(item, itemDef, statDef, talentGrid, itemType);
-
       var createdItem = {
         hash: item.itemHash,
         type: itemType,
         sort: itemSort,
-        tier: (!_.isUndefined(itemDef.tierTypeName) ? itemDef.tierTypeName : 'Common'),
+        tier: itemDef.tierTypeName || 'Common',
         name: itemDef.itemName,
         description: itemDef.itemDescription || '', // Added description for Bounties for now JFLAY2015
         icon: itemDef.icon,
-        notransfer: (itemSort !== 'Postmaster') ? itemDef.nonTransferrable : true,
+        notransfer: (itemSort === 'Postmaster' || itemDef.nonTransferrable),
         id: item.itemInstanceId,
         equipped: item.isEquipped,
         bucket: itemDef.bucketTypeHash,
@@ -422,14 +440,10 @@
         complete: item.isGridComplete,
         amount: item.stackSize,
         primStat: item.primaryStat,
-        stats: stats,
-        quality: getQualityRating(stats, item.primaryStat, itemType, itemDef.itemName),
         // "perks" are the two or so talent grid items that are "featured" for an
         // item in its popup in the game. We don't currently use these.
         //perks: item.perks,
         equipRequiredLevel: item.equipRequiredLevel,
-        talentGrid: talentGrid,
-        objectives: buildObjectives(item, objectiveDef, itemDef),
         maxStackSize: (itemDef.maxStackSize > 0) ? itemDef.maxStackSize : 1,
         // 0: titan, 1: hunter, 2: warlock, 3: any
         classType: itemDef.classType,
@@ -444,6 +458,27 @@
         classified: itemDef.classified
       };
       createdItem.index = createItemIndex(createdItem);
+
+      try {
+        createdItem.talentGrid = buildTalentGrid(item, talentDefs, progressDefs, perkDefs);
+      } catch(e) {
+        console.error("Error building talent grid for " + createdItem.name, item, itemDef);
+      }
+      try {
+        createdItem.stats = buildStats(item, itemDef, statDef, createdItem.talentGrid, itemType);
+      } catch(e) {
+        console.error("Error building stats for " + createdItem.name, item, itemDef);
+      }
+      try {
+        createdItem.objectives = buildObjectives(item, objectiveDef, itemDef);
+      } catch(e) {
+        console.error("Error building objectives for " + createdItem.name, item, itemDef);
+      }
+      try {
+        createdItem.quality = getQualityRating(stats, item.primaryStat, itemType, itemDef.itemName);
+      } catch(e) {
+        console.error("Error building quality rateing for " + createdItem.name, item, itemDef);
+      }
 
       // More objectives properties
       if (createdItem.objectives) {
@@ -749,22 +784,27 @@
     }
 
     function buildStats(item, itemDef, statDef, grid, type) {
-      if (!item.stats || !item.stats.length) {
+      if (!item.stats || !item.stats.length || !itemDef.stats) {
         return undefined;
       }
 
-      var armorNodes, activeArmorNode;
-      if(grid && grid.nodes && item.primaryStat.statHash === 3897883278) {
+      var armorNodes = [];
+      var activeArmorNode;
+      if (grid && grid.nodes && item.primaryStat.statHash === 3897883278) {
         armorNodes = _.filter(grid.nodes, function(node) {
           return _.contains(['Increase Intellect', 'Increase Discipline', 'Increase Strength'], node.name); //[1034209669, 1263323987, 193091484]
         });
-        if(armorNodes) {
-          activeArmorNode = _.findWhere(armorNodes, {activated: true}) || {hash: 0};
+        if (armorNodes) {
+          activeArmorNode = _.findWhere(armorNodes, {activated: true}) || { hash: 0 };
         }
       }
 
       return _.sortBy(_.compact(_.map(itemDef.stats, function(stat) {
         var def = statDef[stat.statHash];
+        if (!def) {
+          return undefined;
+        }
+
         var name = def.statName;
         if (name === 'Aim assistance') {
           name = 'Aim Assist';
@@ -800,13 +840,13 @@
         var base = val;
         var bonus = 0;
 
-        if(item.primaryStat.statHash === 3897883278) {
-          if((name === 'Intellect' && _.find(armorNodes, {name: 'Increase Intellect'})) ||
-             (name === 'Discipline' && _.find(armorNodes, {name: 'Increase Discipline'})) ||
-             (name === 'Strength' && _.find(armorNodes, {name: 'Increase Strength'}))) {
+        if (item.primaryStat.statHash === 3897883278) {
+          if ((name === 'Intellect' && _.find(armorNodes, { name: 'Increase Intellect' })) ||
+             (name === 'Discipline' && _.find(armorNodes, { name: 'Increase Discipline' })) ||
+             (name === 'Strength' && _.find(armorNodes, { name: 'Increase Strength' }))) {
             bonus = getBonus(item.primaryStat.value, type);
 
-            if(activeArmorNode &&
+            if (activeArmorNode &&
                (name === 'Intellect' && activeArmorNode.name === 'Increase Intellect') ||
                (name === 'Discipline' && activeArmorNode.name === 'Increase Discipline') ||
                (name === 'Strength' && activeArmorNode.name === 'Increase Strength')) {
@@ -842,7 +882,12 @@
         .then(function(args) {
           var result = [];
           _.each(items, function (item) {
-            var createdItem = processSingleItem.apply(undefined, args.concat(item));
+            var createdItem = null;
+            try {
+              createdItem = processSingleItem.apply(undefined, args.concat(item));
+            } catch(e) {
+              console.error("Error processing item", item, e);
+            }
             if (createdItem !== null) {
               createdItem.owner = owner;
               result.push(createdItem);
@@ -908,8 +953,17 @@
       var currentBucket = buckets[item.bucket];
 
       // TODO: time to dig through this code
-      if (currentBucket.bucketName === 'Messages') {
+      if (currentBucket && currentBucket.bucketName === 'Messages') {
         return 'Messages';
+      }
+
+      if (name == 'SRL Record Book') {
+        return 'Missions';
+      }
+
+      // File "Mote of Light" under "Material"
+      if (item.itemHash === 937555249) {
+        return 'Material';
       }
 
       if (def.bucketTypeHash === 3621873013) {
@@ -919,6 +973,14 @@
       if (def.bucketTypeHash === 2422292810) {
         if (item.location !== 4)
           return null;
+      }
+
+      if (type.indexOf('Mystery Bag') != -1) {
+        return 'Consumable';
+      }
+
+      if (type.indexOf('Junk') != -1) {
+        return 'Consumable';
       }
 
       if (def.bucketTypeHash === 375726501) {
@@ -1075,6 +1137,10 @@
         }
       }
 
+      if (["Vehicle Upgrade", 'Junk', 'Mystery Bag'].indexOf(type) != -1) {
+        return "Consumable";
+      }
+
       if (typeObj.general !== '') {
         return typeObj;
       }
@@ -1083,7 +1149,7 @@
         return "Messages";
       }
 
-      if (["Vehicle Upgrade"].indexOf(type) != -1) {
+      if (["Vehicle Upgrade", 'Junk', 'Mystery Bag'].indexOf(type) != -1) {
         return "Consumable";
       }