+++ app/scripts/store/dimStoreItems.directive.js
@@ -121,7 +121,7 @@
       }
     };
     vm.onDrop = function(id, $event, equip) {
-      vm.moveDroppedItem(angular.element('#' + id).scope().item, equip, $event);
+      vm.moveDroppedItem(angular.element('#' + id).scope().item, equip, $event, hovering);
       hovering = false;
       dragHelp.classList.remove('drag-dwell-activated');
       $timeout.cancel(dragTimer);
@@ -304,7 +304,7 @@
       }
     };
 
-    vm.moveDroppedItem = dimActionQueue.wrap(function(item, equip, $event) {
+    vm.moveDroppedItem = dimActionQueue.wrap(function(item, equip, $event, hovering) {
       var target = vm.store;
 
       if (item.notransfer && item.owner !== target.id) {