+++ CHANGELOG.md
@@ -1,10 +1,30 @@
 # Next
 
+* Bring back the infusion dialog as an Infusion Fuel Finder. It doesn't do as much as it used to, but now it's optimized for quickly finding eligable infusion items.
+* Fix a bug where hovering over a drop zone with a consumable/material stack and waiting for the message to turn green still wouldn't trigger the partial move dialog.
+
+# 3.5.4
+
+* We won't try to equip an item that is too high-level for your character when dequipping items.
+* Fix a regression where subclasses wouldn't show up in Loadouts. They're still there, they just show up now!
+* Fixed another bug that could prevent item popups from showing up.
+* The vault can now be up to 12 items wide.
+* Sterling Treasure, Junk Items, and SLR Record Book added to DIM.
+* Manifest file updated.
+
+# 3.5.3
+
+* Fixed a bug that would prevent the loading of DIM if Spark of Light was in the postmaster.
+* Fixed a bug that prevented the Xur dialog from rendering.
+
+# 3.5.2
+
 * Fix a bug where item details popups would show above the header.
 * Fix showing Sterling Treasures in Messages.
 * Better error handling when Bungie.net is down.
 * Fix a bug where having items in the postmaster would confuse moves of the same item elsewhere.
 * Fix a bug where item comparisons no longer worked.
+* Added support for the classified shader "Walkabout".
 
 # 3.5.1